requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
sentence-transformers==2.3.1
numpy==1.26.3
prometheus-client==0.19.0
python-dotenv==1.0.0
pyyaml==6.0.1
websockets==12.0
